Split the authorization logic out of the policy evaluation engine
